Moved worker.hpp to component dir.
