+++ test/net_ip/detail/udp_entity_io_test.cpp
@@ -33,7 +33,7 @@
 
 #include "net_ip/detail/udp_entity_io.hpp"
 
-#include "net_ip/worker.hpp"
+#include "net_ip/component/worker.hpp"
 #include "net_ip/endpoints_resolver.hpp"
 
 #include "../test/net_ip/detail/shared_utility_test.hpp"
